The computation of the hashCode is correct, because it is consistent with the implementation of equals(). The hashCode must be different, when two objects are not equal. But the hashcode is allowed to be identical, even if the two objects are different. For example a variable with name "a" and condition "b" and another variable with name "b" and condition "a".
The relevant question would be, whether equals() should focus on name
and conditionid. There could be additional entries in the installer.xml
with same name and same conditionid, but different value. This variable
would be different, but because it is conflicting (which one should be
the winning definition?) this case should be avoided. So defining them
as "equal" and rejecting the second one while reading the installer.xml
would be a good solution.