Avoid NPE when conditionid is not set
The conditionid can be null, so we have to handle this case in equals()
and hashCode().

Because there is a Constructor without arguments the name could be also
null. But in all places using this constructor, the name is set
immediatly after creating the instance. So name==null is not relevant.

The Constructor DynamicVariable() should be replaced with
DynamicVariable(String name), because a variable without a name is not
possible. But this could break implementations outside this project.